move mocks package to internal/
The package should not be used by external packages.
