+++ internal/mocks/consulhealth.go
